+++ feature/home/src/main/java/org/sopt/home/viewholder/HomeClipViewHolder.kt
@@ -8,10 +8,10 @@ import org.sopt.home.databinding.ItemHomeClipBinding
 
 class HomeClipViewHolder(
   private val binding: ItemHomeClipBinding,
-  private val onClickItemClip: (ClipDummy) -> Unit,
-  private val onClickItemClip2: () -> Unit,
+  private val onClickClip: (ClipDummy) -> Unit,
+  private val onClickEmptyClip: () -> Unit,
 
-) : RecyclerView.ViewHolder(binding.root) {
+  ) : RecyclerView.ViewHolder(binding.root) {
   fun onBind(data: ClipDummy?, position: Int) {
     if (position == 0) {
       binding.ivHomeClip.setImageResource(org.sopt.mainfeature.R.drawable.ic_clip_all_24)
@@ -21,7 +21,7 @@ class HomeClipViewHolder(
         clItemClip.isGone = true
         clItemClipEmpty.isVisible = true
         root.setOnClickListener {
-          onClickItemClip2()
+          onClickEmptyClip()
         }
       }
       return
@@ -30,7 +30,7 @@ class HomeClipViewHolder(
       tvItemClipTitle.text = data.title
       tvItemClipCount.text = data.count.toString() + "개"
       root.setOnClickListener {
-        onClickItemClip(data)
+        onClickClip(data)
       }
     }
   }
